+++ include/BDSStackingAction.hh
@@ -23,6 +23,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "G4ClassificationOfNewTrack.hh"
 #include "G4UserStackingAction.hh"
 
+class BDSGlobalConstants;
 class G4Track;
 
 /**
@@ -32,7 +33,7 @@ class G4Track;
 class BDSStackingAction: public G4UserStackingAction
 {
 public:
-  BDSStackingAction();
+  BDSStackingAction(const BDSGlobalConstants* globals);
   virtual ~BDSStackingAction();
 
   /// Decide whether to kill tracks if they're neutrinos or we're killing all secondaries. Note
@@ -43,10 +44,12 @@ public:
   virtual void PrepareNewEvent(); ///< We don't do anything here.
 
 private:
-  G4bool killNeutrinos; ///< Local copy of whether to kill neutrinos for tracking efficiency.
-  G4bool stopSecondaries; ///< Whether particles with parentID > 0 will be killed.
-  G4bool stopTracks;
-  const G4long  maxTracksPerEvent; ///< Maximum number of tracks before start killing.
+  /// Force use of supplied constructor.
+  BDSStackingAction() = delete;
+
+  G4bool killNeutrinos;     ///< Local copy of whether to kill neutrinos for tracking efficiency.
+  G4bool stopSecondaries;   ///< Whether particles with parentID > 0 will be killed.
+  G4long maxTracksPerEvent; ///< Maximum number of tracks before start killing.
  };
 
 #endif