+++ parser/optionsBase.h
@@ -202,7 +202,6 @@ namespace GMAD
     double   minimumEpsilonStep;
     double   maximumEpsilonStep;
     double   deltaOneStep;
-    bool     stopTracks;    
     bool     stopSecondaries;
     bool     killNeutrinos;
     double   minimumRadiusOfCurvature; ///< Minimum allowed radius of curvature.